Завершили проект. Рекомендации  в конце видео по дополнительной прокачке в PHP

+++ inc/fetch.php
@@ -1,5 +1,26 @@
 <?php 
 include 'db.php';
+
+// Make order
+if ($_POST['user_id']) {
+  $user_id = $_POST['user_id'];
+  $query = "SELECT id,qty,price FROM `cart` WHERE user_id={$user_id}"; 
+  $res = mysqli_query($con,$query);
+  $data = mysqli_fetch_all($res,MYSQLI_ASSOC);
+  foreach ($data as $product) {
+      $total = $product['qty']*$product['price'];
+      $query = "INSERT INTO `orders` SET product_id={$product['id']}, user_id={$user_id},qty={$product['qty']},total={$total}";
+      $res = mysqli_query($con,$query);
+      if ($res) {
+          $query = "DELETE FROM `cart` WHERE id={$product['id']}";
+          $res = mysqli_query($con,$query);
+      }
+  }
+  die();
+}
+
+
+
 $id = file_get_contents('php://input');
 $query = "SELECT `qty` FROM `cart` WHERE id={$id}";
 $res = mysqli_query($con, $query);
@@ -11,7 +32,7 @@ if ($data['qty'] > 1) {
   $res = mysqli_query($con, $query);
   $data = mysqli_fetch_assoc($res);
   echo json_encode($data);
-} else {
+} else if ($data['qty'] <= 1) {
   $query = "DELETE FROM `cart` WHERE id={$id}";
   mysqli_query($con, $query);
   $query = "SELECT SUM(qty*price) AS total FROM `cart`";
